ENH: Updates for ITK 5.4.0
Update CI to build against v5.4.0

Also, improve how the `itk` python package version dependency is
specified. Uses `== 5.4.*` to indicated compatibility with 5.4.0, 5.4.1,
5.4.2, etc., but not 5.5.0, etc. There is generally binary
incompatibilty between minor versions. While it is possible that there
could be binary incompatibility between patch releases, it is more
likely that they will be compatible. This flexibilty helps ensure
that remote module packages are not assumed invalid when a new patch
version of `itk` is released. Addresses #142